Rename the submodule containing the quick_fault_detector function to pipeline.py to prevent shadowing the module in the root __init__.py file. Also updates the CLI tests to patch the objects directly.
