+++ energy_fault_detector/quick_fault_detection/__init__.py
@@ -1,2 +1,2 @@
 
-from .quick_fault_detector import quick_fault_detector
+from .pipeline import quick_fault_detector
