+++ energy_fault_detector/quick_fault_detection/pipeline.py
